fix: Update tests and make black happy

+++ tests/transforms/test_remove_global_phases.py
@@ -133,8 +133,8 @@ def test_differentiability_tensorflow():
     dev = qml.device("default.qubit", wires=3)
     original_qnode = qml.QNode(original_qfunc, device=dev)
 
-    phi1 = tf.Variable(0.25)
-    phi2 = tf.Variable(-0.6)
+    phi1 = tf.Variable(0.25, dtype=tf.float64)
+    phi2 = tf.Variable(-0.6, dtype=tf.float64)
     with tf.GradientTape() as tape:
         transformed_qnode = remove_global_phases(original_qnode)(phi1, phi2)
     grad1, grad2 = tape.jacobian(transformed_qnode, (phi1, phi2))
